a program that prints the first 50 Fibonacci numbers

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,7 +8,7 @@
 
 int main(void)
 {
-	int i = 0, n = 1, m = 2, t;
+	long int i = 0, n = 1, m = 2, t;
 
 	printf("1");
 	while (i < 50)
